// src/server/model/iOtherCharges.ts
/**
 * 
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface IOtherCharges { 
    value?: number;
    type?: IOtherCharges.TypeEnum;
    purchaseAccountLedgerId?: number;
    purchaseAccountLedgerName?: string;
    ledgerName?: string;
    ledgerId?: number;
    typeInt?: number;
    discount?: boolean;
    name?: string;
    description?: string;
    modificationdate?: Date;
    impCompanyGuid?: string;
    productUserName?: string;
    productUserId?: number;
    logInfo?: string;
    importedId?: number;
    creationdate?: Date;
    id?: number;
    jacksontype?: string;
}
export namespace IOtherCharges {
    export type TypeEnum = 'FIXED' | 'PERCENT';
    export const TypeEnum = {
        FIXED: 'FIXED' as TypeEnum,
        PERCENT: 'PERCENT' as TypeEnum
    };
}